Release v1.11.0

+++ dongtai-core/src/main/java/io/dongtai/iast/core/utils/HashCode.java
@@ -4,7 +4,7 @@ package io.dongtai.iast.core.utils;
  * @author dongzhiyong@huoxian.cn
  */
 public class HashCode {
-    public static int calc(Object obj) {
+    public static long calc(Object obj) {
         if (obj instanceof String) {
             return ((String) obj).hashCode();
         } else {
